+++ priv/repo/migrations/20221019104919_add_timeout_for_usw_punishments_and_infraction_types.exs
@@ -1,0 +1,10 @@
+defmodule Bolt.Repo.Migrations.AddTimeoutForUswPunishmentsAndInfractionTypes do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+
+  def up do
+    # Please don't question the alternative capitalizations here.
+    execute("ALTER TYPE infraction_type ADD VALUE 'timeout';")
+    execute("ALTER TYPE punishment_type ADD VALUE 'TIMEOUT';")
+  end
+end